LAB-1-From-Problem-to-Approach
Introduction
The aim of these labs is to reinforce the concepts that we discuss in each module's videos. These labs will revolve around the use case of food recipes, and together, we will walk through the process that data scientists usually follow when trying to solve a problem. Let's get started!

In this lab, we will start learning about the data science methodology, and focus on the Business Understanding and the Analytic Approach stages.